+++ client/src/views/QuestionView.vue
@@ -40,8 +40,8 @@
 </template>
 
 <script setup>
-import QuestionNav from '../components/mainquestion/QuestionNav.vue'
-import ChoiceBoard from '../components/mainquestion/ChoiceBoard.vue'
+import QuestionNav from '../components/recommend/QuestionNav.vue'
+import ChoiceBoard from '../components/recommend/ChoiceBoard.vue'
 import HomeButton from '@/components/common/HomeButton.vue'
 import ProfileButton from '@/components/common/ProfileButton.vue'
 import { useRoute, useRouter } from 'vue-router'
